fix/20777/quest-card-refactoring renamed & replaced cardQuest, cardsQuest
