+++ plugins/injectComponents.js
@@ -25,8 +25,8 @@ import FilesUploader from '~/components/ui/FilesUploader';
 import FilesPreview from '~/components/ui/FilesPreview';
 import EmptyData from '~/components/app/info/emptyData';
 import EmployeeCard from '~/components/app/pages/common/employeeCard';
-import QuestCards from '~/components/app/pages/common/quests';
-import QuestCard from '~/components/app/pages/common/questCard';
+import CardsQuest from '~/components/app/cardsQuest';
+import CardQuest from '~/components/app/cardQuest';
 import FiltersPanel from '~/components/app/panels/filters';
 import SearchWithMap from '~/components/app/SearchWithMap';
 import ItemRating from '~/components/app/info/item-rating';
@@ -61,5 +61,5 @@ Vue.component('filters-panel', FiltersPanel);
 Vue.component('search-with-map', SearchWithMap);
 Vue.component('item-rating', ItemRating);
 Vue.component('MessagesList', MessagesList);
-Vue.component('quest-cards', QuestCards);
-Vue.component('quest-card', QuestCard);
+Vue.component('cards-quest', CardsQuest);
+Vue.component('card-quest', CardQuest);
